pyproject.toml
[project]
name = "nonebot-plugin-moyu"
version = "0.4.0"
description = "View your moyu calendar in NoneBot2"
authors = [
    {name = "Akirami", email = "Akiramiaya@outlook.com"},
]
license-expression = "MIT"
dependencies = ["nonebot2>=2.0.0b2", "nonebot-adapter-onebot>=2.0.0b1", "nonebot-plugin-apscheduler>=0.1.2", "httpx[http2]>=0.23.0", "aiofiles>=0.8.0"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/A-kirami/nonebot-plugin-moyu"
Repository = "https://github.com/A-kirami/nonebot-plugin-moyu"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
